backend/pyproject.toml
[project]
name = "daoyoucode"
version = "0.1.0"
description = "新一代AI编程助手"
authors = [
    {name = "daoyoucode Team", email = "team@daoyoucode.dev"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "gitpython>=3.1.40",
    "tree-sitter>=0.20.0",
    "watchdog>=3.0.0",
    # LLM模块依赖
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "jinja2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

[project.scripts]
daoyoucode = "daoyoucode.cli.main:cli"
dyc = "daoyoucode.cli.main:cli"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
